requirements.txt
flask
flask-cors
numpy
tensorflow
Pillow
requests
matplotlib
seaborn
python-dotenv
google-generativeai
